Adding method to login by Facebook and check login status
Button on login screen to login by facebook. When the button is clicked,
opens a new activity of the Facebook, requesting users permissions.
Then the user is redirect to MainActivity, and there is an function to
check the login status, if user is not authenticated, then is redirected to
the login page.

Removing the menu of the action bar of the Login and SignUp activity,
leaving just the main page with option to logout.